src/requirements.txt
aniso8601==9.0.1
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
black==24.2.0
boto3==1.34.62
botocore==1.34.62
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.3
defusedxml==0.7.1
Django==5.0.1
django-cors-headers==3.4.0
django-countries==7.5.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-graphql-auth==0.3.7
django-graphql-jwt==0.4.0
django-modelcluster==6.3
django-permissionedforms==0.1
django-phonenumber-field==7.3.0
django-storages==1.14.2
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-taggit==4.0.0
django-treebeard==4.7.1
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==24.2.0
filetype==1.2.0
gevent==23.9.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==21.2.0
html5lib==1.1
idna==3.6
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
l18n==2021.3
mypy==1.8.0
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.32
pillow==10.2.0
pillow_heif==0.15.0
pipdeptree==2.16.1
platformdirs==4.2.0
pluggy==1.4.0
promise==2.3
psycopg==3.1.17
ptvsd==4.3.2
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.8.0
pytest-env==1.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
ruff==0.3.2
s3transfer==0.10.0
sentry-sdk==1.39.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
text-unidecode==1.3
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240311
typing_extensions==4.10.0
urllib3==2.0.7
wagtail==5.2.3
wagtail-factories==4.1.0
wagtail-grapple==0.24.0
wagtail-headless-preview==0.7.0
wagtail-meta-preview==4.0.0
webencodings==0.5.1
whitenoise==6.6.0
Willow==1.6.3
zope.event==5.0
zope.interface==6.2
